Move media interaction observer
